Refactored directories and resolved merge conflict
