+++ include/boost/astronomy/coordinate/coord_sys/coordinate.hpp
@@ -6,12 +6,13 @@
   file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
 
-#ifndef BOOST_ASTRONOMY_COORDINATE_REPRESENTATION_HPP
-#define BOOST_ASTRONOMY_COORDINATE_REPRESENTATION_HPP
 
-#include <boost/astronomy/coordinate/cartesian_representation.hpp>
-#include <boost/astronomy/coordinate/spherical_equatorial_representation.hpp>
-#include <boost/astronomy/coordinate/spherical_representation.hpp>
+#ifndef BOOST_ASTRONOMY_COORDINATE_HPP
+#define BOOST_ASTRONOMY_COORDINATE_HPP
 
-#endif // !BOOST_ASTRONOMY_COORDINATE_REPRESENTATION_HPP
+#include <boost/astronomy/coordinate/rep/representation.hpp>
+#include <boost/astronomy/coordinate/diff/differential.hpp>
+#include <boost/astronomy/coordinate/ref_frame/frame.hpp>
+
+#endif // !BOOST_ASTRONOMY_COORDINATE_HPP
