+++ include/boost/astronomy/coordinate/coord_sys/geocentric.hpp
@@ -7,11 +7,10 @@
   file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
 
-#ifndef BOOST_ASTRONOMY_COORDINATE_CIRS_HPP
-#define BOOST_ASTRONOMY_COORDINATE_CIRS_HPP
+#ifndef BOOST_ASTRONOMY_COORDINATE_GEOCENTRIC_HPP
+#define BOOST_ASTRONOMY_COORDINATE_GEOCENTRIC_HPP
 
-#include <boost/date_time/posix_time/posix_time.hpp>
-#include <boost/astronomy/coordinate/base_equatorial_frame.hpp>
+#include <boost/astronomy/coordinate/ref_frame/base_ecliptic_frame.hpp>
 
 namespace boost { namespace astronomy { namespace coordinate {
 
@@ -19,63 +18,49 @@ template
 <
     typename Representation, typename Differential
 >
-struct cirs : public base_equatorial_frame<Representation, Differential>
+struct geocentric : public base_ecliptic_frame<Representation, Differential>
 {
-
-protected:
-    //!time used to determine the position of earth at the time of observation
-    boost::posix_time::ptime obs_time; 
 
 public:
     //default constructor no initialization
-    cirs() {}
+    geocentric() {}
 
     //!constructs object from another representation object
     template <typename OtherRepresentation>
-    cirs(OtherRepresentation const& representation_data) : base_equatorial_frame
+    geocentric(OtherRepresentation const& representation_data) : base_ecliptic_frame
         <Representation, Differential>(representation_data) {}
 
     //!constructs object from provided components of representation
-    cirs
+    geocentric
     (
-        typename Representation::quantity1 const& dec,
-        typename Representation::quantity2 const& ra,
+        typename Representation::quantity1 const& lat,
+        typename Representation::quantity2 const& lon,
         typename Representation::quantity3 const& distance
-    ) : base_equatorial_frame<Representation, Differential>(dec, ra, distance) {}
+    ) : base_ecliptic_frame<Representation, Differential>(lat, lon, distance) {}
 
     //!constructs object from provided components of representation and differential
-    cirs
+    geocentric
     (
-        typename Representation::quantity1 const& dec,
-        typename Representation::quantity2 const& ra,
+        typename Representation::quantity1 const& lat,
+        typename Representation::quantity2 const& lon,
         typename Representation::quantity3 const& distance,
-        typename Differential::quantity1 const& pm_dec,
-        typename Differential::quantity2 const& pm_ra_cosdec,
+        typename Differential::quantity1 const& pm_lat,
+        typename Differential::quantity2 const& pm_lon_coslat,
         typename Differential::quantity3 const& radial_velocity
-    ) : base_equatorial_frame<Representation, Differential>
-            (dec, ra, distance, pm_dec, pm_ra_cosdec, radial_velocity) {}
+    ) : base_ecliptic_frame<Representation, Differential>
+            (lat, lon, distance, pm_lat, pm_lon_coslat, radial_velocity) {}
 
     //!constructs object from other representation and differential objects
     template <typename OtherRepresentation, typename OtherDifferential>
-    cirs
+    geocentric
     (
         OtherRepresentation const& representation_data,
         OtherDifferential const& differential_data
-    ) : base_equatorial_frame<Representation, Differential>
+    ) : base_ecliptic_frame<Representation, Differential>
             (representation_data, differential_data) {}
-
-    boost::posix_time::ptime get_obs_time() const
-    {
-        return this->obstime;
-    }
-
-    void set_obs_time(boost::posix_time::ptime const& time)
-    {
-        this->obstime = time;
-    }
 };
 
 }}} //namespace boost::astronomy::coordinate
 
-#endif // !BOOST_ASTRONOMY_COORDINATE_CIRS_HPP
+#endif // !BOOST_ASTRONOMY_COORDINATE_GEOCENTRIC_HPP
