+++ include/boost/astronomy/coordinate/diff/differential.hpp
@@ -9,10 +9,10 @@
 #ifndef BOOST_ASTRONOMY_COORDINATE_DIFFERENTIAL_HPP
 #define BOOST_ASTRONOMY_COORDINATE_DIFFERENTIAL_HPP
 
-#include <boost/astronomy/coordinate/cartesian_differential.hpp>
-#include <boost/astronomy/coordinate/spherical_differential.hpp>
-#include <boost/astronomy/coordinate/spherical_equatorial_differential.hpp>
-#include <boost/astronomy/coordinate/spherical_coslat_differential.hpp>
+#include <boost/astronomy/coordinate/diff/cartesian_differential.hpp>
+#include <boost/astronomy/coordinate/diff/spherical_differential.hpp>
+#include <boost/astronomy/coordinate/diff/spherical_equatorial_differential.hpp>
+#include <boost/astronomy/coordinate/diff/spherical_coslat_differential.hpp>
 
 #endif // !BOOST_ASTRONOMY_COORDINATE_DIFFERENTIAL_HPP
