+++ include/boost/astronomy/coordinate/diff/spherical_differential.hpp
@@ -7,8 +7,8 @@
   file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
 
-#ifndef BOOST_ASTRONOMY_COORDINATE_SPHERICAL_EQUATORIAL_DIFFERENTIAL_HPP
-#define BOOST_ASTRONOMY_COORDINATE_SPHERICAL_EQUATORIAL_DIFFERENTIAL_HPP
+#ifndef BOOST_ASTRONOMY_COORDINATE_SPHERICAL_DIFFERENTIAL_HPP
+#define BOOST_ASTRONOMY_COORDINATE_SPHERICAL_DIFFERENTIAL_HPP
 
 #include <tuple>
 #include <type_traits>
@@ -25,26 +25,26 @@
 #include <boost/units/systems/si/plane_angle.hpp>
 
 #include <boost/astronomy/detail/is_base_template_of.hpp>
-#include <boost/astronomy/coordinate/base_differential.hpp>
-#include <boost/astronomy/coordinate/cartesian_differential.hpp>
+#include <boost/astronomy/coordinate/diff/base_differential.hpp>
+#include <boost/astronomy/coordinate/diff/cartesian_differential.hpp>
 
 namespace boost { namespace astronomy { namespace coordinate {
 
 namespace bu = boost::units;
 namespace bg = boost::geometry;
 
-//!Represents the differential in spherical equatorial representation
+//!Represents the differential in spherical representation
 //!Uses three components to represent a differential (dlatitude, dlongitude, ddistance)
 
-template
+template 
 <
     typename CoordinateType = double,
     typename LatQuantity = bu::quantity<bu::si::plane_angle, CoordinateType>,
     typename LonQuantity = bu::quantity<bu::si::plane_angle, CoordinateType>,
-    typename DistQuantity = bu::quantity<bu::si::dimensionless, CoordinateType>
+    typename DistQuantity = bu::quantity<bu::si::dimensionless, CoordinateType>    
 >
-struct spherical_equatorial_differential : public base_differential
-    <3, bg::cs::spherical_equatorial<radian>, CoordinateType>
+struct spherical_differential : public base_differential
+    <3, bg::cs::spherical<radian>, CoordinateType>
 {
     ///@cond INTERNAL
     BOOST_STATIC_ASSERT_MSG(
@@ -63,10 +63,10 @@ public:
     typedef DistQuantity quantity3;
 
     //default constructor no initialization
-    spherical_equatorial_differential() {}
+    spherical_differential() {}
 
     //!constructs object from provided components of differential
-    spherical_equatorial_differential
+    spherical_differential
     (
         LatQuantity const& dlat,
         LonQuantity const& dlon,
@@ -83,7 +83,7 @@ public:
         typename OtherCoordinateSystem,
         typename OtherCoordinateType
     >
-    spherical_equatorial_differential
+    spherical_differential
     (
         bg::model::point
         <
@@ -99,34 +99,34 @@ public:
     }
 
     //copy constructor
-    spherical_equatorial_differential
+    spherical_differential
     (
-        spherical_equatorial_differential
+        spherical_differential
         <
             CoordinateType,
             LatQuantity,
             LonQuantity,
             DistQuantity
-        > const& object
+        > const& other
     )
     {
-        this->diff = object.get_differential();
+        this->diff = other.get_differential();
     }
 
     //!constructs object from any type of differential
     template <typename Differential>
-    spherical_equatorial_differential(Differential const& other)
+    spherical_differential(Differential const& other)
     {
         BOOST_STATIC_ASSERT_MSG((boost::astronomy::detail::is_base_template_of
             <boost::astronomy::coordinate::base_differential, Differential>::value),
             "No constructor found with given argument type");
 
-         BOOST_STATIC_ASSERT_MSG(
+        BOOST_STATIC_ASSERT_MSG(
         ((std::is_same<typename bu::get_dimension<DistQuantity>::type,
         typename bu::get_dimension<typename Differential::quantity3>::type>::value)),
         "Two differentials must have same dimensions");
 
-        auto tempDiff = make_spherical_equatorial_differential(other);
+        auto tempDiff = make_spherical_differential(other);
         bg::model::point
         <
             typename std::conditional
@@ -136,7 +136,7 @@ public:
                 typename Differential::type
             >::type,
             3,
-            bg::cs::spherical_equatorial<radian>
+            bg::cs::spherical<radian>
         > tempPoint;
         
         bg::set<0>(tempPoint,
@@ -151,8 +151,8 @@ public:
             static_cast<DistQuantity>(tempDiff.get_ddist()).value());
 
         this->diff = tempPoint;
-
     }
+
     //! returns the (dlat, dlon, ddistance) in the form of tuple
     std::tuple<LatQuantity, LonQuantity, DistQuantity> get_dlat_dlon_ddist() const
     {
@@ -229,13 +229,13 @@ public:
     <
         typename Addend
     >
-    spherical_equatorial_differential
+    spherical_differential
     <
         CoordinateType,
         LatQuantity,
         LonQuantity,
         DistQuantity
-    >
+    > 
     operator +(Addend const& addend) const
     {
         auto cartesian1 = make_cartesian_differential
@@ -244,13 +244,13 @@ public:
 
         auto temp = cartesian1 + cartesian2;
 
-        spherical_equatorial_differential
+        spherical_differential
         <
             CoordinateType,
             LatQuantity,
             LonQuantity,
             DistQuantity
-        > result = make_spherical_equatorial_differential(temp);
+        > result = make_spherical_differential(temp);
 
         return result;
     }
@@ -262,11 +262,11 @@ public:
     >
     auto operator *(OtherQuantity const& dt) const
     {
-        
-        spherical_equatorial_differential
-            <CoordinateType, LatQuantity, LonQuantity, DistQuantity> temp(this->diff);
 
-        spherical_equatorial_differential
+        spherical_differential<CoordinateType, LatQuantity, LonQuantity, DistQuantity>
+            temp(this->diff);
+
+        spherical_differential
         <
             CoordinateType,
             LatQuantity,
@@ -286,9 +286,10 @@ public:
 
         return product;
     }
-}; //spherical_equatorial_differential
+}; //spherical_differential
 
 //!constructs object from provided components of differential
+//!Each quantity can have different units but with same datatypes
 template
 <
     typename CoordinateType,
@@ -299,21 +300,21 @@ template
     typename Unit2,
     typename Unit3
 >
-spherical_equatorial_differential
+spherical_differential
 <
     CoordinateType,
     LatQuantity<Unit1, CoordinateType>,
     LonQuantity<Unit2, CoordinateType>,
     DistQuantity<Unit3, CoordinateType>
 >
-make_spherical_equatorial_differential
+make_spherical_differential
 (
     LatQuantity<Unit1, CoordinateType> const& dlat,
     LonQuantity<Unit2, CoordinateType> const& dlon,
     DistQuantity<Unit3, CoordinateType> const& ddist
 )
 {
-    return spherical_equatorial_differential
+    return spherical_differential
         <
             CoordinateType,
             LatQuantity<Unit1, CoordinateType>,
@@ -334,16 +335,16 @@ template
     typename LonQuantity,
     typename DistQuantity
 >
-spherical_equatorial_differential
+spherical_differential
 <
     ReturnCoordinateType,
     ReturnLatQuantity,
     ReturnLonQuantity,
     ReturnDistQuantity
 >
-make_spherical_equatorial_differential
+make_spherical_differential
 (
-    spherical_equatorial_differential
+    spherical_differential
     <
         CoordinateType,
         LatQuantity,
@@ -352,7 +353,7 @@ make_spherical_equatorial_differential
     > const& other
 )
 {
-    return make_spherical_equatorial_differential(
+    return make_spherical_differential(
         static_cast<ReturnLatQuantity>(other.get_dlat()),
         static_cast<ReturnLonQuantity>(other.get_dlon()),
         static_cast<ReturnDistQuantity>(other.get_ddist())
@@ -367,16 +368,16 @@ template
     typename LonQuantity,
     typename DistQuantity
 >
-spherical_equatorial_differential
+spherical_differential
 <
     CoordinateType,
     LatQuantity,
     LonQuantity,
     DistQuantity
 >
-make_spherical_equatorial_differential
+make_spherical_differential
 (
-    spherical_equatorial_differential
+    spherical_differential
     <
         CoordinateType,
         LatQuantity,
@@ -385,7 +386,7 @@ make_spherical_equatorial_differential
     > const& other
 )
 {
-    return spherical_equatorial_differential
+    return spherical_differential
         <
             CoordinateType,
             LatQuantity,
@@ -405,14 +406,14 @@ template
     typename OtherCoordinateSystem,
     typename OtherCoordinateType
 >
-spherical_equatorial_differential
+spherical_differential
 <
     CoordinateType,
     LatQuantity,
     LonQuantity,
     DistQuantity
 >
-make_spherical_equatorial_differential
+make_spherical_differential
 (
     bg::model::point
     <
@@ -422,7 +423,7 @@ make_spherical_equatorial_differential
     > const& pointObject
 )
 {
-    return spherical_equatorial_differential
+    return spherical_differential
         <
             CoordinateType,
             LatQuantity,
@@ -436,11 +437,11 @@ template
 <
     typename OtherDifferential
 >
-auto make_spherical_equatorial_differential
+auto make_spherical_differential
 (
     OtherDifferential const& other
 )
-{
+{   
     auto temp = make_cartesian_differential(other);
     typedef decltype(temp) cartesian_type;
 
@@ -457,11 +458,10 @@ auto make_spherical_equatorial_differential
     bg::set<2>(tempPoint, static_cast<typename cartesian_type::quantity1>
         (temp.get_dz()).value());
 
-    bg::model::point<typename cartesian_type::type, 3, bg::cs::spherical_equatorial
-        <radian>> result;
+    bg::model::point<typename cartesian_type::type, 3, bg::cs::spherical<radian>> result;
     bg::transform(tempPoint, result);
 
-    return spherical_equatorial_differential
+    return spherical_differential
         <
             typename cartesian_type::type,
             bu::quantity<bu::si::plane_angle, typename cartesian_type::type>,
@@ -472,6 +472,5 @@ auto make_spherical_equatorial_differential
 
 }}}//namespace boost::astronomy::coordinate
 
-#endif // !BOOST_ASTRONOMY_COORDINATE_SPHERICAL_EQUATORIAL_DIFFERENTIAL_HPP
-
+#endif // !BOOST_ASTRONOMY_COORDINATE_SPHERICAL_DIFFERENTIAL_HPP
