+++ include/boost/astronomy/coordinate/ref_frame/frame.hpp
@@ -9,13 +9,13 @@
 #ifndef BOOST_ASTRONOMY_COORDINATE_FRAME_HPP
 #define BOOST_ASTRONOMY_COORDINATE_FRAME_HPP  
 
-#include <boost/astronomy/coordinate/alt_az.hpp>
-#include <boost/astronomy/coordinate/cirs.hpp>
-#include <boost/astronomy/coordinate/galactic.hpp>
-#include <boost/astronomy/coordinate/geocentric.hpp>
-#include <boost/astronomy/coordinate/heliocentric.hpp>
-#include <boost/astronomy/coordinate/icrs.hpp>
-#include <boost/astronomy/coordinate/supergalactic.hpp>
+#include <boost/astronomy/coordinate/coord_sys/alt_az.hpp>
+#include <boost/astronomy/coordinate/ref_frame/cirs.hpp>
+#include <boost/astronomy/coordinate/coord_sys/galactic.hpp>
+#include <boost/astronomy/coordinate/coord_sys/geocentric.hpp>
+#include <boost/astronomy/coordinate/coord_sys/heliocentric.hpp>
+#include <boost/astronomy/coordinate/ref_frame/icrs.hpp>
+#include <boost/astronomy/coordinate/coord_sys/supergalactic.hpp>
 
 #endif // !BOOST_ASTRONOMY_COORDINATE_FRAME_HPP
