+++ include/boost/astronomy/coordinate/ref_frame/icrs.hpp
@@ -7,10 +7,10 @@
   file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
 
-#ifndef BOOST_ASTRONOMY_COORDINATE_HELIOCENTRIC_HPP
-#define BOOST_ASTRONOMY_COORDINATE_HELIOCENTRIC_HPP
+#ifndef BOOST_ASTRONOMY_COORDINATE_ICRS_HPP
+#define BOOST_ASTRONOMY_COORDINATE_ICRS_HPP
 
-#include <boost/astronomy/coordinate/base_ecliptic_frame.hpp>
+#include <boost/astronomy/coordinate/ref_frame/base_equatorial_frame.hpp>
 
 namespace boost { namespace astronomy { namespace coordinate {
 
@@ -18,49 +18,49 @@ template
 <
     typename Representation, typename Differential
 >
-struct heliocentric : public base_ecliptic_frame<Representation, Differential>
+struct icrs : public base_equatorial_frame<Representation, Differential>
 {
 
 public:
     //default constructor no initialization
-    heliocentric() {}
+    icrs() {}
 
     //!constructs object from another representation object
     template <typename OtherRepresentation>
-    heliocentric(OtherRepresentation const& representation_data) : base_ecliptic_frame
+    icrs(OtherRepresentation const& representation_data) : base_equatorial_frame
         <Representation, Differential>(representation_data) {}
 
     //!constructs object from provided components of representation
-    heliocentric
+    icrs
     (
-        typename Representation::quantity1 const& lat,
-        typename Representation::quantity2 const& lon,
+        typename Representation::quantity1 const& dec,
+        typename Representation::quantity2 const& ra,
         typename Representation::quantity3 const& distance
-    ) : base_ecliptic_frame<Representation, Differential>(lat, lon, distance) {}
+    ) : base_equatorial_frame<Representation, Differential>(dec, ra, distance) {}
 
     //!constructs object from provided components of representation and differential
-    heliocentric
+    icrs
     (
-        typename Representation::quantity1 const& lat,
-        typename Representation::quantity2 const& lon,
+        typename Representation::quantity1 const& dec,
+        typename Representation::quantity2 const& ra,
         typename Representation::quantity3 const& distance,
-        typename Differential::quantity1 const& pm_lat,
-        typename Differential::quantity2 const& pm_lon_coslat,
+        typename Differential::quantity1 const& pm_dec,
+        typename Differential::quantity2 const& pm_ra_cosdec,
         typename Differential::quantity3 const& radial_velocity
-    ) : base_ecliptic_frame<Representation, Differential>
-            (lat, lon, distance, pm_lat, pm_lon_coslat, radial_velocity) {}
+    ) : base_equatorial_frame<Representation, Differential>
+            (dec, ra, distance, pm_dec, pm_ra_cosdec, radial_velocity) {}
 
     //!constructs object from other representation and differential objects
     template <typename OtherRepresentation, typename OtherDifferential>
-    heliocentric
+    icrs
     (
         OtherRepresentation const& representation_data,
         OtherDifferential const& differential_data
-    ) : base_ecliptic_frame<Representation, Differential>
+    ) : base_equatorial_frame<Representation, Differential>
             (representation_data, differential_data) {}
 };
 
 }}} //namespace boost::astronomy::coordinate
 
-#endif // !BOOST_ASTRONOMY_COORDINATE_HELIOCENTRIC_HPP
+#endif // !BOOST_ASTRONOMY_COORDINATE_ICRS_HPP
