+++ include/boost/astronomy/coordinate/rep/spherical_equatorial_representation.hpp
@@ -7,13 +7,13 @@
   file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
 
-#ifndef BOOST_ASTRONOMY_COORDINATE_SPHERICAL_REPRESENTATION_HPP
-#define BOOST_ASTRONOMY_COORDINATE_SPHERICAL_REPRESENTATION_HPP
-
+#ifndef BOOST_ASTRONOMY_COORDINATE_SPHERICAL_EQUATORIAL_REPRESENTATION_HPP
+#define BOOST_ASTRONOMY_COORDINATE_SPHERICAL_EQUATORIAL_REPRESENTATION_HPP
 
 #include <tuple>
 #include <cstddef>
 
+#include <boost/geometry/strategies/strategy_transform.hpp>
 #include <boost/static_assert.hpp>
 #include <boost/geometry/core/cs.hpp>
 #include <boost/geometry/geometries/point.hpp>
@@ -21,11 +21,12 @@
 #include <boost/geometry/algorithms/equals.hpp>
 #include <boost/units/physical_dimensions/plane_angle.hpp>
 #include <boost/units/systems/si/plane_angle.hpp>
+#include <boost/units/get_dimension.hpp>
 #include <boost/units/systems/si/dimensionless.hpp>
 
 #include <boost/astronomy/detail/is_base_template_of.hpp>
-#include <boost/astronomy/coordinate/base_representation.hpp>
-#include <boost/astronomy/coordinate/cartesian_representation.hpp>
+#include <boost/astronomy/coordinate/rep/base_representation.hpp>
+#include <boost/astronomy/coordinate/rep/cartesian_representation.hpp>
 
 
 namespace boost { namespace astronomy { namespace coordinate {
@@ -34,22 +35,14 @@ namespace bu = boost::units;
 namespace bg = boost::geometry;
 
 
-//!Represents the coordinate in spherical representation
+//!Represents the coordinate in spherical equatorial representation
 //!Uses three components to represent a point/vector (latitude, longitude, distance)
 /*!
-\brief Spherical (polar) coordinate system, in degree or in radian
-\details Defines the spherical coordinate system where points are
-    defined in two angles
-    and an optional radius usually known as r, theta, phi
-\par Coordinates:
-- coordinate 0:
-    0 <= phi < 2pi is the angle between the positive x-axis and the
-        line from the origin to the P projected onto the xy-plane.
-- coordinate 1:
-    0 <= theta <= pi is the angle between the positive z-axis and the
-        line formed between the origin and P.
-- coordinate 2 (if specified):
-    r >= 0 is the distance from the origin to a given point P.
+\brief Spherical equatorial coordinate system, in degree or in radian
+\details This one resembles the geographic coordinate system, and has latitude
+    up from zero at the equator, to 90 at the pole
+    (opposite to the spherical(polar) coordinate system).
+    Used in astronomy and in GIS (but there is also the geographic)
 
 \see http://en.wikipedia.org/wiki/Spherical_coordinates
 \ingroup cs
@@ -61,8 +54,8 @@ template
     typename LonQuantity = bu::quantity<bu::si::plane_angle, CoordinateType>,
     typename DistQuantity = bu::quantity<bu::si::dimensionless, CoordinateType>
 >
-struct spherical_representation : public base_representation
-    <3, bg::cs::spherical<radian>, CoordinateType>
+struct spherical_equatorial_representation : public base_representation
+    <3, bg::cs::spherical_equatorial<radian>, CoordinateType>
 {
     ///@cond INTERNAL
     BOOST_STATIC_ASSERT_MSG(
@@ -81,10 +74,10 @@ public:
     typedef DistQuantity quantity3;
 
     //default constructor no initialization
-    spherical_representation() {}
+    spherical_equatorial_representation(){}
 
     //!constructs object from provided value of coordinates (latitude, longitude, distance)
-    spherical_representation
+    spherical_equatorial_representation
     (
         LatQuantity const& lat,
         LonQuantity const& lon,
@@ -101,7 +94,7 @@ public:
         typename OtherCoordinateSystem,
         typename OtherCoordinateType
     >
-    spherical_representation
+    spherical_equatorial_representation
     (
         bg::model::point
         <
@@ -117,22 +110,23 @@ public:
     }
 
     //copy constructor
-    spherical_representation
+    spherical_equatorial_representation
     (
-        spherical_representation
+        spherical_equatorial_representation
         <
             CoordinateType,
             LatQuantity,
             LonQuantity,
             DistQuantity
-        > const& other)
+        > const& object
+    )
     {
-        this->point = other.get_point();
+        this->point = object.get_point();
     }
 
     //!constructs object from any type of representation
     template <typename Representation>
-    spherical_representation(Representation const& other)
+    spherical_equatorial_representation(Representation const& other)
     {
         BOOST_STATIC_ASSERT_MSG((boost::astronomy::detail::is_base_template_of
             <boost::astronomy::coordinate::base_representation, Representation>::value),
@@ -143,7 +137,7 @@ public:
         typename bu::get_dimension<typename Representation::quantity3>::type>::value)),
         "Two representations must have same dimensions");
 
-        auto tempRep = make_spherical_representation(other);
+        auto tempRep = make_spherical_equatorial_representation(other);
         bg::model::point
         <
             typename std::conditional
@@ -153,7 +147,7 @@ public:
                 typename Representation::type
             >::type,
             3,
-            bg::cs::spherical<radian>
+            bg::cs::spherical_equatorial<radian>
         > tempPoint;
         
         bg::set<0>(tempPoint,
@@ -168,8 +162,8 @@ public:
             static_cast<DistQuantity>(tempRep.get_dist()).value());
 
         this->point = tempPoint;
-    }
 
+    }
     //! returns the (lat, lon, distance) in the form of tuple
     std::tuple<LatQuantity, LonQuantity, DistQuantity> get_lat_lon_dist() const
     {
@@ -241,8 +235,9 @@ public:
         bg::set<2>(this->point, distance.value());
     }
 
+    //!operator for addition of representation
     template<typename Addend>
-    spherical_representation
+    spherical_equatorial_representation
     <
         CoordinateType,
         LatQuantity,
@@ -252,25 +247,24 @@ public:
     operator +(Addend const& addend) const
     {
 
-
         auto cartesian1 = make_cartesian_representation
             <CoordinateType, DistQuantity, DistQuantity, DistQuantity>(this->point);
         auto cartesian2 = make_cartesian_representation(addend);
 
         auto temp = cartesian1 + cartesian2;
 
-        spherical_representation
+        spherical_equatorial_representation
         <
             CoordinateType,
             LatQuantity,
             LonQuantity,
             DistQuantity
-        > result = make_spherical_representation(temp);
+        > result = make_spherical_equatorial_representation(temp);
 
         return result;
     }
 
-}; //spherical_representation
+}; //spherical_equatorial_representation
 
 
 //!Constructs object from provided quantities
@@ -285,20 +279,21 @@ template
     typename Unit2,
     typename Unit3
 >
-spherical_representation
+spherical_equatorial_representation
 <
     CoordinateType,
     LatQuantity<Unit1, CoordinateType>,
     LonQuantity<Unit2, CoordinateType>,
     DistQuantity<Unit3, CoordinateType>
-> make_spherical_representation
+>
+make_spherical_equatorial_representation
 (
     LatQuantity<Unit1, CoordinateType> const& lat,
     LonQuantity<Unit2, CoordinateType> const& lon,
     DistQuantity<Unit3, CoordinateType> const& dist
 )
 {
-    return spherical_representation
+    return spherical_equatorial_representation
         <
             CoordinateType,
             LatQuantity<Unit1, CoordinateType>,
@@ -307,7 +302,7 @@ spherical_representation
         >(lat, lon, dist);
 }
 
-//!Convert current quantities of spherical_representation to new quantities. 
+//!constructs object from provided components of representation with different units
 template
 <
     typename ReturnCoordinateType,
@@ -319,16 +314,16 @@ template
     typename LonQuantity,
     typename DistQuantity
 >
-spherical_representation
+spherical_equatorial_representation
 <
     ReturnCoordinateType,
     ReturnLatQuantity,
     ReturnLonQuantity,
     ReturnDistQuantity
 >
-make_spherical_representation
+make_spherical_equatorial_representation
 (
-    spherical_representation
+    spherical_equatorial_representation
     <
         CoordinateType,
         LatQuantity,
@@ -337,14 +332,14 @@ make_spherical_representation
     > const& other
 )
 {
-    return make_spherical_representation(
+    return make_spherical_equatorial_representation(
         static_cast<ReturnLatQuantity>(other.get_lat()),
         static_cast<ReturnLonQuantity>(other.get_lon()),
         static_cast<ReturnDistQuantity>(other.get_dist())
     );
 }
 
-//!Create copy of spherical_representation
+//!constructs object from provided representation
 template
 <
     typename CoordinateType,
@@ -352,10 +347,16 @@ template
     typename LonQuantity,
     typename DistQuantity
 >
-spherical_representation<CoordinateType, LatQuantity, LonQuantity, DistQuantity>
-make_spherical_representation
+spherical_equatorial_representation
+<
+    CoordinateType,
+    LatQuantity,
+    LonQuantity,
+    DistQuantity
+>
+make_spherical_equatorial_representation
 (
-    spherical_representation
+    spherical_equatorial_representation
     <
         CoordinateType,
         LatQuantity,
@@ -364,21 +365,19 @@ make_spherical_representation
     > const& other
 )
 {
-    return spherical_representation
+    return spherical_equatorial_representation
         <
-        CoordinateType,
-        LatQuantity,
-        LonQuantity,
-        DistQuantity
+            CoordinateType,
+            LatQuantity,
+            LonQuantity,
+            DistQuantity
         >(other);
 }
 
-//!Create spherical_representation from boost::geometry::point
-//!Quantity types are to be specifed explicitly or angle is considered to be in radians
-//!distance is taken dimensionless by default if not specified 
+//!constructs object from boost::geometry::model::point object
 template
 <
-    typename CoordinateType=double,
+    typename CoordinateType = double,
     typename LatQuantity = bu::quantity<bu::si::plane_angle, CoordinateType>,
     typename LonQuantity = bu::quantity<bu::si::plane_angle, CoordinateType>,
     typename DistQuantity = bu::quantity<bu::si::dimensionless, CoordinateType>,
@@ -386,31 +385,38 @@ template
     typename OtherCoordinateSystem,
     typename OtherCoordinateType
 >
-spherical_representation<CoordinateType, LatQuantity, LonQuantity, DistQuantity>
-make_spherical_representation
+spherical_equatorial_representation
+<
+    CoordinateType,
+    LatQuantity,
+    LonQuantity,
+    DistQuantity
+>
+make_spherical_equatorial_representation
 (
     bg::model::point
     <
-    OtherCoordinateType,
-    OtherDimensionCount,
-    OtherCoordinateSystem
-    > const& pointObject)
+        OtherCoordinateType,
+        OtherDimensionCount,
+        OtherCoordinateSystem
+    > const& pointObject
+)
 {
-    return spherical_representation
+    return spherical_equatorial_representation
         <
-        CoordinateType,
-        LatQuantity,
-        LonQuantity,
-        DistQuantity
+            CoordinateType,
+            LatQuantity,
+            LonQuantity,
+            DistQuantity
         >(pointObject);
 }
 
-//!Create cartesian_representation from other type of representations
-template
+//!constructs object from any type of representation
+template 
 <
     typename OtherRepresentation
 >
-auto make_spherical_representation
+auto make_spherical_equatorial_representation
 (
     OtherRepresentation const& other
 )
@@ -418,26 +424,24 @@ auto make_spherical_representation
     auto temp = make_cartesian_representation(other);
     typedef decltype(temp) cartesian_type;
 
-    bg::model::point<typename cartesian_type::type, 3, bg::cs::cartesian> tempPoint;
-    bg::model::point<typename cartesian_type::type, 3, bg::cs::spherical<radian>> result;
+    bg::model::point
+        <
+            typename cartesian_type::type,
+            3,
+            bg::cs::cartesian
+        > tempPoint;
 
     bg::set<0>(tempPoint, temp.get_x().value());
-    bg::set<1>
-    (
-        tempPoint,
-        static_cast<typename cartesian_type::quantity1>
-        (temp.get_y()).value()
-    );
-    bg::set<2>
-    (
-        tempPoint,
-        static_cast<typename cartesian_type::quantity1>
-        (temp.get_z()).value()
-    );
+    bg::set<1>(tempPoint, static_cast<typename cartesian_type::quantity1>
+        (temp.get_y()).value());
+    bg::set<2>(tempPoint, static_cast<typename cartesian_type::quantity1>
+        (temp.get_z()).value());
 
+    bg::model::point<typename cartesian_type::type, 3, bg::cs::spherical_equatorial
+        <radian>> result;
     bg::transform(tempPoint, result);
 
-    return spherical_representation
+    return spherical_equatorial_representation
         <
             typename cartesian_type::type,
             bu::quantity<bu::si::plane_angle, typename cartesian_type::type>,
@@ -445,7 +449,8 @@ auto make_spherical_representation
             typename cartesian_type::quantity1
         >(result);
 }
-
+        
 }}} //namespace boost::astronomy::coordinate
 
-#endif // !BOOST_ASTRONOMY_COORDINATE_SPHERICAL_REPRESENTATION_HPP
+#endif // !BOOST_ASTRONOMY_COORDINATE_SPHERICAL_EQUATORIAL__REPRESENTATION_HPP
+
